+++ Classes/Controller/SubscribeController.php
@@ -1,140 +1,135 @@
 <?php
 namespace XQueue\Typo3MaileonIntegration\Controller;
 
+use Exception;
 use Psr\Http\Message\ResponseInterface;
+use TYPO3\CMS\Core\Http\HtmlResponse;
+use TYPO3\CMS\Core\Log\LogManager;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Extbase\Persistence\Exception\InvalidQueryException;
 use TYPO3\CMS\Extbase\Utility\DebuggerUtility;
+use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
 use XQueue\Typo3MaileonIntegration\Domain\Model\Subscription;
-use de\xqueue\maileon\api\client\contacts\SynchronizationMode;
-use de\xqueue\maileon\api\client\contacts\Permission;
-use de\xqueue\maileon\api\client\contacts\ContactsService;
-use de\xqueue\maileon\api\client\contacts\Contact;
+use XQueue\Typo3MaileonIntegration\Domain\Repository\XQHbSendRepository;
+use XQueue\Typo3MaileonIntegration\Services\FormProcessingService;
+use XQueue\Typo3MaileonIntegration\Services\HeartBeatService;
 
 /**
  * SubscribeController
  */
 class SubscribeController extends ActionController
 {
+    protected FormProcessingService $formProcessingService;
+
+    protected XQHbSendRepository $xqHbSendRepository;
+
+    protected HeartBeatService $heartBeatService;
+
+    public function initializeAction(): void
+    {
+        $logManager = GeneralUtility::makeInstance(LogManager::class);
+
+        $this->formProcessingService = new FormProcessingService($this->settings);
+        $this->xqHbSendRepository = GeneralUtility::makeInstance(XQHbSendRepository::class);
+        $this->heartBeatService = new HeartBeatService($this->settings, $logManager);
+    }
+    public function showAction(): ResponseInterface
+    {
+        $this->view->setTemplatePathAndFilename(
+            'EXT:typo3_maileon_integration/Resources/Private/Templates/Subscribe/Subscribe.html'
+        );
+
+        $this->assignFormSettingsToView();
+
+        // Render the view and return the response
+        $content = $this->view->render();
+        return new HtmlResponse($content);
+    }
     /**
      * Subscribe action
      *
-     * @param Subscription|null $data
      * @return ResponseInterface
      */
-    public function subscribeAction(?Subscription $data = null): ResponseInterface
+    public function subscribeAction(): ResponseInterface
     {
-        if ($data != null) {
-            // validate data
-            $isValid = true;
+        $submittedData = $this->request->getArguments();
+        $errors = $this->formProcessingService->validateData($submittedData);
+        $this->assignFormSettingsToView();
 
-            if (false == $data->isApproval()) {
-                $isValid = false;
-                $this->view->assign("error_approval", true);
-                return $this->htmlResponse();
-            }
+        if (!empty($errors)) {
+            // Return errors and form values back to the view
+            $this->view->assign('errors', $errors);
+            $this->view->assign('formValues', $submittedData);
 
-            if (false == $data->isPrivacy()) {
-                $isValid = false;
-                $this->view->assign("error_privacy", true);
-                return $this->htmlResponse();
-            }
-
-            if (true == $isValid) {
-                // send data via maileon api
-                $contactsService = $this->getContactsService();
-
-                // Check if contact exists in Maileon
-                $getContactByEmail = $contactsService->getContactByEmail($data->getEmail());
-
-                $newContact = $this->createNewContact($data);
-
-                if ($getContactByEmail->isSuccess()) {
-                    $existingContact = $getContactByEmail->getResult();
-
-                    // Check permission of existing customer
-                    if (Permission::$NONE == $existingContact->permission) {
-                        // Contact does exist without permission - create with doi mailing
-                        $response = $contactsService->createContact(
-                            $newContact,
-                            SynchronizationMode::$UPDATE,
-                            'Typo3',
-                            'subscriptionForm',
-                            true,
-                            ($this->settings["targetPermission"] == 5),
-                            $this->settings["doiMailingKey"]
-                        );
-                    } else {
-                        // Contact does exist with permission - just update customer data
-                        $response = $contactsService->createContact(
-                            $newContact,
-                            SynchronizationMode::$UPDATE,
-                            'Typo3',
-                            'subscriptionForm'
-                        );
-                    }
-                } else {
-                    // Contact does not exist - create with doi mailing
-                    $response = $contactsService->createContact(
-                        $newContact,
-                        SynchronizationMode::$UPDATE,
-                        'Typo3',
-                        'subscriptionForm',
-                        true,
-                        ($this->settings["targetPermission"] == 5),
-                        $this->settings["doiMailingKey"]
-                    );
-                }
-
-                if ($response->isSuccess()) {
-                    // redirect to result page
-                    if ($this->settings["targetSuccess"]) {
-                        return $this->redirectByPid($this->settings["targetSuccess"]);
-                    }
-                } else {
-                    if ($this->settings["debug"] == 1) {
-                        // show debug information
-                        DebuggerUtility::var_dump($response->getBodyData());
-                    } else {
-                        // redirect to error page
-                        if ($this->settings["targetError"]) {
-                            return $this->redirectByPid($this->settings["targetError"]);
-                        }
-                    }
-                }
-            } else {
-                // return to form and show error
-                $this->view->assign("data", $data);
-                return $this->htmlResponse();
-            }
+            $content = $this->view->render();
+            return new HtmlResponse($content);
         }
 
-        return $this->htmlResponse();
+        try {
+            $this->handleHB();
+            $response = $this->formProcessingService->trySubscribeContact($submittedData);
+
+            if ($response->isSuccess()) {
+                // redirect to result page
+                if ($this->settings["targetSuccess"]) {
+                    return $this->redirectByPid($this->settings["targetSuccess"]);
+                }
+
+                $this->addFlashMessage(
+                    LocalizationUtility::translate('subscribe.success', 'typo3_maileon_integration')
+                );
+                $this->redirect('show');
+            } else {
+                if ($this->settings["debug"] == 1) {
+                    // show debug information
+                    DebuggerUtility::var_dump($response->getBodyData());
+                } else {
+                    // redirect to error page
+                    if ($this->settings["targetError"]) {
+                        return $this->redirectByPid($this->settings["targetError"]);
+                    }
+
+                    $this->addFlashMessage(
+                        LocalizationUtility::translate('submit.failed', 'typo3_maileon_integration')
+                    );
+                }
+            }
+        } catch (Exception $e) {
+            $this->view->assign('errors', $e->getMessage());
+        }
+
+        $content = $this->view->render();
+        return new HtmlResponse($content);
     }
 
     /**
      * Unsubscribe action
-     *
-     * @param Subscription|null $data
-     * @return void
      */
     public function unsubscribeAction(?Subscription $data = null): ResponseInterface
     {
         if ($data != null) {
             // validate data
             if (null != $data->getEmail()) {
-                // send data via maileon api
-                $contactsService = $this->getContactsService();
-                $response = $contactsService->unsubscribeContactByEmail($data->getEmail());
+                $response = $this->formProcessingService->tryUnsubscribeContact($data->getEmail());
 
                 // redirect to result page
                 if ($response->isSuccess()) {
                     if ($this->settings["targetSuccess"]) {
                         return $this->redirectByPid($this->settings["targetSuccess"]);
                     }
+
+                    $this->addFlashMessage(
+                        LocalizationUtility::translate('unsubscribe.success', 'typo3_maileon_integration')
+                    );
                 } else {
                     if ($this->settings["targetError"]) {
                         return $this->redirectByPid($this->settings["targetError"]);
                     }
+
+                    $this->addFlashMessage(
+                        LocalizationUtility::translate('submit.failed', 'typo3_maileon_integration')
+                    );
                 }
             }
         }
@@ -144,123 +139,52 @@ class SubscribeController extends ActionController
 
     /**
      * Redirect by page id
-     *
-     * @param integer $pid
-     * @return ResponseInterface
      */
-    private function redirectByPid(int $pid): ResponseInterface
+    protected function redirectByPid(int $pid): ResponseInterface
     {
         $uriBuilder = $this->uriBuilder;
         $uriBuilder->reset();
         $uri = $uriBuilder
-        ->setTargetPageUid($pid)
-        ->build();
+            ->setTargetPageUid($pid)
+            ->build();
 
         return $this->redirectToUri($uri);
     }
 
-    /**
-     * Returns contacts service from maileon api
-     *
-     * @return ContactsService
-     */
-    private function getContactsService()
+    protected function assignFormSettingsToView(): void
     {
-        // Set the global configuration for accessing the REST-API
-        $config = array(
-          "BASE_URI" => "https://api.maileon.com/1.0",
-          "API_KEY" => $this->settings["apiKey"],
-          "TIMEOUT" => 30,
-          "DEBUG" => "false" // NEVER enable on production
-        );
+        $standardFields = $this->settings['subscribeForm']['standardFields'] ?? [];
+        $customFields = $this->settings['subscribeForm']['customFields'] ?? [];
+        $privacyPolicyUrl = $this->settings['privacyPolicyUrl'] ?? '#';
 
-        $contactsService = new ContactsService($config);
-        $contactsService->setDebug(false);
+        $activeStandardFields = array_filter($standardFields, function ($field) {
+            return $field['active'] ?? false;
+        });
 
-        return $contactsService;
+        $activeCustomFields = array_filter($customFields, function ($field) {
+            return $field['active'] ?? false;
+        });
+
+        $this->view->assign('privacyPolicyUrl', $privacyPolicyUrl);
+        $this->view->assign('standardFields', $activeStandardFields);
+        $this->view->assign('customFields', $activeCustomFields);
     }
 
     /**
-     * Create Contact obj
-     *
-     * @param Subscription $data
-     * @return Contact
+     * @throws InvalidQueryException
      */
-    private function createNewContact(Subscription $data)
+    protected function handleHB(): void
     {
-        $newContact = new Contact();
-        $newContact->anonymous = false;
+        $task = $this->xqHbSendRepository->findByTask('maileon_hb');
 
-        $newContact->email = $data->getEmail();
-
-        // Standard and custom fields
-        $standard_fields = $this->extractMaileonStandardFields($data);
-        $custom_fields = $this->extractMaileonCustomFields($data);
-
-        $newContact->standard_fields = $standard_fields;
-        $newContact->custom_fields = $custom_fields;
-
-        if (!empty($newContact->custom_fields)) {
-            $this->checkAndCreateCustomFields($newContact->custom_fields);
+        if (empty($task)) {
+            $this->heartBeatService->executeHB();
+            $this->xqHbSendRepository->updateLastExecution('maileon_hb');
         }
 
-        return $newContact;
-    }
-
-    /**
-     * Get array for standard fields
-     *
-     * @param Subscription $data
-     * @return array
-     */
-    private function extractMaileonStandardFields(Subscription $data)
-    {
-        return [
-            "SALUTATION" => $data->getSalutation(),
-            "FIRSTNAME" => $data->getFirstname(),
-            "LASTNAME" => $data->getLastname(),
-            "ORGANIZATION" => $data->getOrganization()
-        ];
-    }
-
-    /**
-     * Get array for custom fields
-     *
-     * @param Subscription $data
-     * @return array
-     */
-    private function extractMaileonCustomFields(Subscription $data)
-    {
-        return [
-            "Position" => $data->getPosition(),
-            "SubscriptionNumber" => $data->getSubscriptionnumber(),
-            "Typo3_created" => true
-        ];
-    }
-
-    /**
-     * Check the custom fields exist at Maileon. If not create it.
-     *
-     * @param array $customFields
-     * @return boolean
-     */
-    private function checkAndCreateCustomFields(array $customFields): bool
-    {
-        $contactsService = $this->getContactsService();
-
-        $customFieldsResponse = $contactsService->getCustomFields();
-        $customFieldsFromMaileon = $customFieldsResponse->getResult();
-
-        if (!array_key_exists('Typo3_created', $customFieldsFromMaileon->custom_fields)) {
-            $contactsService->createCustomField('Typo3_created', 'boolean');
+        if (! $this->xqHbSendRepository->hasTaskRunToday('maileon_hb')) {
+            $this->heartBeatService->executeHB();
+            $this->xqHbSendRepository->updateLastExecution('maileon_hb');
         }
-
-        foreach ($customFields as $fieldName => $fieldValue) {
-            if (!array_key_exists($fieldName, $customFieldsFromMaileon->custom_fields)) {
-                $contactsService->createCustomField($fieldName, 'string');
-            }
-        }
-
-        return true;
     }
 }