+++ Classes/Settings/Settings.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace XQueue\Typo3MaileonIntegration\Settings;
+
+final class Settings
+{
+    public const XSIC_ID = '10023';
+    public const XSIC_CHECKSUM = 'dbWIvEDI69UXf22mpRxqctkH';
+    public const XSIC_URL = 'https://integrations.maileon.com/xsic/tx.php';
+}